backend/requirements.txt
fastapi
uvicorn
motor
pymongo
bcrypt
python-jose
SpeechRecognition
pyaudio
pyttsx3
gTTS
pydantic
pydantic-settings
email-validator
python-dotenv
